Display main chromatogram for a sample
- take chromatogram functions from dash-version and adjust them to work
without dash (e.g. returning the graph as a JSON string so it can be
used by JavaScript)
- adjust route in views.py accordingly
- change output of access_loaded_data() to return proper dictionary
rather than the original JSON-format of it (to adhere to DRY-principle)
- empty variables.py to only keep those variables that are already
needed (others will be added again, when appropriate)
- move some functions from sample_table.py to dashboard_functions.py
because they're also used for the chromatogram
