+++ src/fermo/app_utils/variables.py
@@ -1,27 +1,3 @@
-style_data_table = {
-    'color': 'black',
-    'backgroundColor': 'white',
-    'font-size': '12px',
-    'textAlign': 'left',
-    'whiteSpace': 'normal',
-    'height': 'auto',
-}
-
-style_data_cond_table = [{
-    'if': {'row_index': 'odd'},
-    'backgroundColor': 'rgb(220, 220, 220)',
-}]
-
-style_header_table = {
-    'backgroundColor': '#116789',
-    'color': 'white',
-    'fontWeight': 'bold',
-    'textAlign': 'left',
-    'font-size': '15px',
-    'whiteSpace': 'normal',
-    'height': 'auto',
-}
-
 # Colorblind-safe palette
 # (inspired by Fig.1 in https://personal.sron.nl/~pault/)
 color_dict = {
@@ -44,100 +20,4 @@ color_dict = {
     'very_light_grey': 'rgba(187,187,187,0.3)',
     'black': 'rgba(0,0,0,1)',
     'light_black': 'rgba(0,0,0,0.5)',
-}
-
-stylesheet_cytoscape = [
-    {
-        'selector': '.selected',
-        'style': {
-            'background-color': color_dict['light_blue'],
-            'border-width': 4,
-            'border-color': color_dict['blue'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.selected_unique_sample',
-        'style': {
-            'background-color': color_dict['light_blue'],
-            'border-width': 4,
-            'border-color': color_dict['purple'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.selected_unique_group',
-        'style': {
-            'background-color': color_dict['light_blue'],
-            'border-width': 4,
-            'border-color': color_dict['black'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.sample',
-        'style': {
-            'background-color': color_dict['light_red'],
-            'border-width': 4,
-            'border-color': color_dict['red'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.sample_unique_sample',
-        'style': {
-            'background-color': color_dict['light_red'],
-            'border-width': 4,
-            'border-color': color_dict['purple'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.sample_unique_group',
-        'style': {
-            'background-color': color_dict['light_red'],
-            'border-width': 4,
-            'border-color': color_dict['black'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.default',
-        'style': {
-            'background-color': color_dict['light_grey'],
-            'border-width': 4,
-            'border-color': color_dict['grey'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.default_unique_sample',
-        'style': {
-            'background-color': color_dict['light_grey'],
-            'border-width': 4,
-            'border-color': color_dict['purple'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '.default_unique_group',
-        'style': {
-            'background-color': color_dict['light_grey'],
-            'border-width': 4,
-            'border-color': color_dict['black'],
-            'label': 'data(label)',
-        }
-    }, {
-        'selector': '[weight > 0.8]',
-        'style': {
-            'width': 2,
-        }
-    }, {
-        'selector': '[weight > 0.85]',
-        'style': {
-            'width': 3,
-        }
-    }, {
-        'selector': '[weight > 0.90]',
-        'style': {
-            'width': 5,
-        }
-    }, {
-        'selector': '[weight > 0.95]',
-        'style': {
-            'width': 7,
-        }
-    },
-]
+    }
